refactor url_base into conftest.py; move non-a11y test into test_playwright.py
