+++ tests/test_a11y.py
@@ -1,10 +1,5 @@
 """Using Axe-core, scan the Kitchen Sink pages for accessibility violations."""
 
-import re
-import time
-from http.client import HTTPConnection
-from pathlib import Path
-from subprocess import PIPE, Popen
 from urllib.parse import urljoin
 
 import pytest
@@ -25,50 +20,6 @@ from playwright.sync_api import Page, expect  # noqa: E402
 # *not* mean that it will be generally usable by a broad range of disabled
 # people. It just means that page is free of common testable accessibility
 # pitfalls.
-
-path_repo = Path(__file__).parent.parent
-path_docs_build = path_repo / "docs" / "_build" / "html"
-
-
-@pytest.fixture(scope="module")
-def url_base():
-    """Start local server on built docs and return the localhost URL as the base URL."""
-    # Use a port that is not commonly used during development or else you will
-    # force the developer to stop running their dev server in order to run the
-    # tests.
-    port = "8213"
-    host = "localhost"
-    url = f"http://{host}:{port}"
-
-    # Try starting the server
-    process = Popen(
-        ["python", "-m", "http.server", port, "--directory", path_docs_build],
-        stdout=PIPE,
-    )
-
-    # Try connecting to the server
-    retries = 5
-    while retries > 0:
-        conn = HTTPConnection(host, port)
-        try:
-            conn.request("HEAD", "/")
-            response = conn.getresponse()
-            if response is not None:
-                yield url
-                break
-        except ConnectionRefusedError:
-            time.sleep(1)
-            retries -= 1
-
-    # If the code above never yields a URL, then we were never able to connect
-    # to the server and retries == 0.
-    if not retries:
-        raise RuntimeError("Failed to start http server in 5 seconds")
-    else:
-        # Otherwise the server started and this fixture is done now and we clean
-        # up by stopping the server.
-        process.terminate()
-        process.wait()
 
 
 def filter_ignored_violations(violations, url_pathname):
@@ -228,24 +179,6 @@ def test_axe_core(
     assert len(filtered_violations) == 0, format_violations(filtered_violations)
 
 
-def test_version_switcher_highlighting(page: Page, url_base: str) -> None:
-    """This isn't an a11y test, but needs a served site for Javascript to inject the version menu."""
-    page.goto(url=url_base)
-    # no need to include_hidden here ↓↓↓, we just need to get the active version name
-    button = page.get_by_role("button").filter(has_text="dev")
-    active_version_name = button.get_attribute("data-active-version-name")
-    # here we do include_hidden, so sidebar & topbar menus should each have a matching entry:
-    entries = page.get_by_role("option", include_hidden=True).filter(
-        has_text=active_version_name
-    )
-    assert entries.count() == 2
-    # make sure they're highlighted
-    for entry in entries.all():
-        light_mode = "rgb(10, 125, 145)"  # pst-color-primary
-        # dark_mode = "rgb(63, 177, 197)"
-        expect(entry).to_have_css("color", light_mode)
-
-
 @pytest.mark.a11y
 def test_code_block_tab_stop(page: Page, url_base: str) -> None:
     """Code blocks that have scrollable content should be tab stops."""
@@ -308,8 +241,13 @@ def test_notebook_ipywidget_output_tab_stop(page: Page, url_base: str) -> None:
     # (tabIndex = 0)
     assert ipywidget.evaluate("el => el.tabIndex") == 0
 
-@pytest.mark.a11y
-def test_breadcrumb_expansion(page: Page, url_base: str) -> None:
-    page.goto(urljoin(url_base, "community/practices/merge.html"))
-    expect(page.get_by_label("Breadcrumb").get_by_role("list")).to_contain_text("Merge and review policy")
 
+def test_breadcrumb_expansion(page: Page, url_base: str) -> None:
+    """Foo."""
+    # page.goto(urljoin(url_base, "community/practices/merge.html"))
+    # expect(page.get_by_label("Breadcrumb").get_by_role("list")).to_contain_text("Merge and review policy")
+    page.set_viewport_size({"width": 1440, "height": 720})
+    page.goto(urljoin(url_base, "community/topics/config.html"))
+    expect(page.get_by_label("Breadcrumb").get_by_role("list")).to_contain_text(
+        "Update Sphinx configuration during the build"
+    )
